Fade-out working! Overlay not
We have come so far! When you put in a name and date and press submit, the correct zodiac sign is determined, the loading symbol spins, and the content fades out. However, the overlay does not yet fade in, and I'm not sure how to add the image and description to it.