Added Bottom Icons
I added the little icons to represent each zodiac across the bottom of the screen. They should be set up so clicking on them brings up the overlay for the correct zodiac, but for some reason they are producing an error.
The error of not being able to get a clear overlay after closing the window 2+ times still persists.
Otherwise, everything is good.